fix: add GetPositionRange function to all parser nodes
The EndPos that participle provides is actually the start position of
the next token, even if that token is in a different node, which makes
it hard to format errors properly.

The GetPositionRange function returns a start and end position where the
start position is just what participle provides but the end position
is calculated by finding the end position of the last token in _this_ node.