chore: gql input to modelmap value (#164)
Context: As part of the API we auto generate for the user, there is code to handle Create actions, and it needs to receive incoming request arguments and convert their types to be suitable for a) inserting into a database column, b) to return from the request handler. These differ in a few cases - at the moment, specifically those fields that are to do with time.

Code to look at:

the entry point is: /runtime/actions/create.go
that provides context, but the thing this PR is really about is...
/runtime/actions/model.go/toMap() and its tests.
Caveats:
The code in create.go that writes to the database is: a statement of intent and compiles, but hasn't been run yet, (requires test suite design) and probably doesn't work yet. That is for another PR.
I suspect either "we", or "I" have not got our approach to time-oriented values quite right yet. And consequently there are some tests commented out. But I'd like to get the rest of the relevant code looked at before trying to solve those problems.